+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Change Log
+<!-- # Change Log
 
 ## [1.10.0] 2020-05-07
 ### Bug fixing
@@ -354,4 +354,4 @@ ajv                        6.9.1   →   6.10.0
 ## [1.0.0] 2018-01-29
 ### Original Release
 - Added Material-UI as base framework
-- Added design from Material Dashboard by Creative Tim
+- Added design from Material Dashboard by Creative Tim -->